Rename "scrapeconfigs" Package to "parcascrapeconfig"
The "scrapeconfigs" package was renamed to "parcascrapeconfig" to follow
the naming of the controller.
